solved some compilation warnings

+++ modules/devices/robot/obstacle_avoidance/obstacle_avoidance.c
@@ -172,7 +172,7 @@ point_t * oa_get_path(void)
 
 void oa_dump(void)
 {
-	uint16_t i,j;
+	uint8_t i,j;
 	poly_t *poly;
 	point_t *pt;
 
@@ -184,7 +184,7 @@ void oa_dump(void)
 		printf_P(PSTR("poly #%d\r\n"), i);
 		for (j=0; j<poly->l; j++) {
 			pt = &poly->pts[j];
-			printf_P(PSTR("  pt #%d (%"PRIi32",%"PRIi32")\r\n"), j, pt->x, pt->y);
+			printf_P(PSTR("  pt #%d (%2.2f,%2.2f)\r\n"), j, pt->x, pt->y);
 		}
 	}
 }
